+++ bin/itemplate.js
@@ -110,6 +110,7 @@ return /******/ (function(modules) { // webpackBootstrap
 	        open: '{%',
 	        close: '%}'
 	    },
+	    emptyString: true,
 	    MAP: {
 	        '&': '&amp;',
 	        '<': '&lt;',
@@ -611,12 +612,24 @@ return /******/ (function(modules) { // webpackBootstrap
 
 	function decodeAccessory(string) {
 	    var regex = new RegExp(_options.accessory.open + '|' + _options.accessory.close, 'g');
-	    var code; var isStatic = true;
+	    var code;
+	    var isStatic = true, openStub, closeStub;
 
 	    code = string.split(regex).map(function (piece, i) {
+	        openStub = '';
+	        closeStub = '';
+
 	        if (i % 2) {
 	            isStatic = false;
-	            return ' + ' + piece.trim() + ' + ';
+	            piece = piece.trim();
+	            if (_options.emptyString) { // undefined as empty string
+	                if (piece.indexOf(' ') !== -1) {
+	                    openStub = '(';
+	                    closeStub = ')';
+	                }
+	                return ' + ' + openStub + piece + closeStub + '||\'\' + ';
+	            } else
+	                return ' + ' + piece + ' + ';
 	        } else {
 	            return JSON.stringify(piece);
 	        }
@@ -625,13 +638,13 @@ return /******/ (function(modules) { // webpackBootstrap
 	    // micro-optimizations (remove appending empty strings)
 	    code = code.replace(/^"" \+ | \+ ""$/g, '').replace(/ \+ "" \+ /g, ' + ');
 
-	    return { value: code, isStatic: isStatic};
+	    return {value: code, isStatic: isStatic};
 	}
 
 	function formatText(tag, text) {
 	    return text
 	        .replace(_options.BREAK_LINE, ((_options.textSaveTags.indexOf(tag) !== -1) ? '\n' : ' '))
-	        .replace(/&#(\d+);/g, function(match, dec) {
+	        .replace(/&#(\d+);/g, function (match, dec) {
 	            return String.fromCharCode(dec);
 	        })
 	        .trim();
@@ -750,7 +763,6 @@ return /******/ (function(modules) { // webpackBootstrap
 
 	Builder.prototype.write = function (command) {
 	    var tag;
-	    console.log(command)
 	    switch (command.type) {
 	        case Mode.Tag:
 	            tag = command.name.replace('/', empty);
@@ -822,10 +834,8 @@ return /******/ (function(modules) { // webpackBootstrap
 	            if (_fnName) // plain function as string
 	                resultFn = 'function ' + _fnName + '(' + _options.parameterName + ', lib, helpers){'
 	                    + fn + stack.join('') + '}';
-	            else{ // plain function
-	            console.log(stack.join(''))
+	            else // plain function
 	                resultFn = new Function(_options.parameterName, 'lib', 'helpers', fn + stack.join(''));
-	                }
 	        }
 
 	        return resultFn;